:rotating_light: Removed circular dependency
When compiling my project with Rollup.js, it throws this warning:

```
(!) Circular dependency: node_modules/bezier-js/lib/bezier.js
  -> node_modules/bezier-js/lib/utils.js
  -> node_modules/bezier-js/lib/bezier.js
```

So bezier needs utils, which needs bezier.

I had a look, and the only method in utils that needs bezier is
the `makeline()` method.

I've removed it from utils.js, and instead added it to utils once
imported in bezier.js.

This avoid the `require` statement, but I guess it's kinda hackish.

I'm certain there's a better way to fix this, but I wanted to (try
to) be a bit more helpful than open an issue. Hence this PR.

Note that this does not make any functional changes to the code.
Nor is the issue that it fixes a *real* problem. It would just be
nice to avoid the circular dependency and suppress that warning.